+++ tests/test_labelbuddy.py
@@ -2,6 +2,7 @@ import json
 import math
 from unittest.mock import Mock
 
+import pandas as pd
 import pytest
 
 from pubget import (
@@ -33,7 +34,7 @@ def test_bad_batch_size(tmp_path):
             )
 
 
-@pytest.mark.parametrize("batch_size", [3, None, 1000])
+@pytest.mark.parametrize("batch_size", [3, None, 1000, "n_docs"])
 def test_make_labelbuddy_documents(
     batch_size, tmp_path, entrez_mock, monkeypatch
 ):
@@ -43,14 +44,16 @@ def test_make_labelbuddy_documents(
     )
     articles_dir, _ = _articles.extract_articles(download_dir)
     data_dir, _ = _data_extraction.extract_data_to_csv(articles_dir)
+    n_articles = json.loads(
+        articles_dir.joinpath("info.json").read_text("utf-8")
+    )["n_articles"]
+    if batch_size == "n_docs":
+        batch_size = n_articles
     labelbuddy_dir, code = _labelbuddy.make_labelbuddy_documents(
         data_dir, batch_size=batch_size
     )
     assert code == ExitCode.COMPLETED
     assert labelbuddy_dir.name == "subset_allArticles_labelbuddyData"
-    n_articles = json.loads(
-        articles_dir.joinpath("info.json").read_text("utf-8")
-    )["n_articles"]
     expected_batch_size = (
         n_articles if batch_size is None else min(batch_size, n_articles)
     )
@@ -63,6 +66,21 @@ def test_make_labelbuddy_documents(
         docs = [json.loads(doc_json) for doc_json in f]
     assert len(docs) == expected_batch_size
     assert all("Body\n The text of" in d["text"] for d in docs)
+    _check_batch_info(labelbuddy_dir)
+
+
+def _check_batch_info(labelbuddy_dir):
+    batch_info = pd.read_csv(labelbuddy_dir.joinpath("batch_info.csv"))
+    for file_name, file_contents in batch_info.groupby("file_name"):
+        docs = [
+            json.loads(doc_line)
+            for doc_line in labelbuddy_dir.joinpath(file_name)
+            .read_text("utf-8")
+            .strip()
+            .split("\n")
+        ]
+        for pmcid, _, line in file_contents.values:
+            assert pmcid == docs[line]["metadata"]["pmcid"]
 
 
 @pytest.mark.parametrize(